api/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai>=0.5.0
python-dotenv